Update the matrix module
Replace the Matrix module with the Matrix4x4 module for working with 4x4
matrices. Update the Matrix references throughout the code. Remove the
Vector and Point modules and all references to these modules throughout
the code. Update the Quaternion module to define a struct instead of
using a tuple to represent a quaternion.